refactor folder structure in frontend
