+++ frontend/src/app/mod.rs
@@ -1,0 +1,22 @@
+mod index;
+
+use yew::{Component, Context, Html, html};
+
+use index::Index;
+
+pub struct App;
+
+impl Component for App {
+    type Message = ();
+    type Properties = ();
+
+    fn create(ctx: &Context<Self>) -> Self {
+        Self
+    }
+
+    fn view(&self, ctx: &Context<Self>) -> Html {
+        html! {
+            <Index/>
+        }
+    }
+}
