+++ frontend/src/main.rs
@@ -1,5 +1,3 @@
-use gloo_timers::callback::Timeout;
-use linked_hash_set::LinkedHashSet;
 use tracing::Level;
 use tracing_subscriber::fmt::time::UtcTime;
 use tracing_subscriber::fmt::writer::MakeWriterExt;
@@ -7,74 +5,13 @@ use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::util::SubscriberInitExt;
 use yew::prelude::*;
 
-use link_form::LinkForm;
-use message_box::MessageBox;
-
-use crate::message_box::Message;
+use crate::app::App;
 use crate::util::fetch_server_config;
 
-mod link_input;
-mod advanced_mode;
-mod toggle;
-mod expiration_mode;
 mod util;
-mod duration_input;
-mod link_form;
-mod message_box;
-mod link_config;
-
-
-enum AppMessage {
-    AddMessage(Message),
-    ClearMessage(Message),
-}
-
-struct App {
-    messages: LinkedHashSet<Message>,
-}
-
-impl Component for App {
-    type Message = AppMessage;
-    type Properties = ();
-
-    fn create(_: &Context<Self>) -> Self {
-        Self {
-            messages: LinkedHashSet::new(),
-        }
-    }
-
-    fn update(&mut self, ctx: &Context<Self>, msg: Self::Message) -> bool {
-        match msg {
-            AppMessage::AddMessage(m) => {
-                self.messages.insert(m.clone());
-
-                let callback = ctx.link().callback(|m| AppMessage::ClearMessage(m));
-                let timeout = Timeout::new(5_000, move || {
-                    callback.emit(m);
-                });
-
-                timeout.forget();
-            },
-            AppMessage::ClearMessage(m) => { self.messages.remove(&m); },
-        }
-
-        true
-    }
-
-    fn view(&self, ctx: &Context<Self>) -> Html {
-        let callback = ctx.link().callback(|m| {
-            AppMessage::AddMessage(m)
-        });
-
-        html! {
-            <>
-                <MessageBox messages={ self.messages.clone() }/>
-                <h1>{ "[WIP] Link Shortener" }</h1>
-                <LinkForm callback={ callback }/>
-            </>
-        }
-    }
-}
+mod components;
+mod app;
+mod types;
 
 fn setup_tracing_subscriber() {
     // done with consts because of https://github.com/rust-lang/rust/issues/15701
@@ -108,13 +45,8 @@ fn setup_tracing_subscriber() {
             - this includes redoing the message format or rephrasing error messages
         - migrate to yew 0.21.0
     TODO less important:
-        - checkout if the form name coupling can be made more concise
-            - maybe use contexts
-            - or by using node_refs and dropping the name entirely (currently sounds the best)
         - decide between console_error_panic_hook and color_eyre and decide if any of these is needed at all
         - add a footer to the page for stuff like "about", "github source", ...
-        - organize the components in different modules and maybe spilt or merge a few files
-            - maybe after atomic design logic, but that would need a bigger restructure
         - should callbacks always be assigned in the view() method or is there another way (static?), that's more optimized
         - do not use the App component as as a middleman but talk directly to message box
         - currently the logic for "disabling" the submit functionality just change the button type.
@@ -128,6 +60,7 @@ fn main() {
     setup_tracing_subscriber();
     // used to fetch the server config in the beginning
     fetch_server_config();
+
     // panic::set_hook(Box::new(console_error_panic_hook::hook));
 
     yew::Renderer::<App>::new().render();
